Add unit test for Steady State Ph1D reactor
