+++ run_tests.py
@@ -1,7 +1,7 @@
 from asali.reactors.batch import BatchReactor
 from asali.reactors.cstr import CstrReactor
 from asali.reactors.het1d import Heterogeneous1DReactor
-from asali.reactors.ph1d import PseudoHomogeneous1DReactor
+from asali.reactors.ph1d_steady_state import SteadyStatePseudoHomogeneous1DReactor
 from asali.utils.input_parser import InputParser
 from tests.basic_unit_test import BasicUnitTest
 from tests.reactor_unit_test import ReactorUnitTest
@@ -69,13 +69,13 @@ def cstr_reactor(is_local):
     print(" ")
 
 
-def pseudohomogeneous_reactor(is_local):
+def steady_state_pseudohomogeneous_reactor(is_local):
     """
-    Run unit test for PseudoHomogeneous1DReactor
+    Run unit test for SteadyStatePseudoHomogeneous1DReactor
     :param is_local: Bool that shows if the test is performed locally or not
     :return:
     """
-    pfr = PseudoHomogeneous1DReactor('tests/H2-O2-Rh.yaml', 'gas', 'Rh_surface')
+    pfr = SteadyStatePseudoHomogeneous1DReactor('tests/H2-O2-Rh.yaml', 'gas', 'Rh_surface')
     ut = ReactorUnitTest(pfr, "reactors.json", is_local)
     ut.check_all()
     print(" ")
@@ -110,7 +110,7 @@ if __name__ == "__main__":
         cantera_file_converter()
         batch_reactor(args.is_local)
         cstr_reactor(args.is_local)
-        pseudohomogeneous_reactor(args.is_local)
+        steady_state_pseudohomogeneous_reactor(args.is_local)
         heterogeneous_reactor(args.is_local)
     elif class_to_check == "unitconverter":
         unit_converter()
@@ -118,8 +118,8 @@ if __name__ == "__main__":
         batch_reactor(args.is_local)
     elif class_to_check == "cstrreactor":
         cstr_reactor(args.is_local)
-    elif class_to_check == "pseudohomogeneous1dreactor":
-        pseudohomogeneous_reactor(args.is_local)
+    elif class_to_check == "steadystatepseudohomogeneous1dreactor":
+        steady_state_pseudohomogeneous_reactor(args.is_local) #TODO prepare for steady state and transient
     elif class_to_check == "heterogeneous1dreactor":
         heterogeneous_reactor(args.is_local)
     elif class_to_check == "canterafileconverter":
