--- tests/PseudoHomogeneous1DReactor/get_coverage.asali
@@ -1,10 +1,0 @@
-1.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
-9.743547839999999738e-01 1.085892719999999905e-06 1.096796809999999942e-02 2.498343510000000005e-04 1.443166870000000060e-02
-9.743556540000000155e-01 1.085893520000000055e-06 1.096797440000000026e-02 2.498347150000000097e-04 1.443169220000000037e-02
-9.743561649999999963e-01 1.085894110000000077e-06 1.096797790000000015e-02 2.498349539999999763e-04 1.443170790000000009e-02
-9.743557729999999806e-01 1.085893269999999995e-06 1.096797260000000052e-02 2.498348930000000032e-04 1.443170319999999979e-02
-9.743551660000000503e-01 1.085892609999999930e-06 1.096796759999999968e-02 2.498346620000000047e-04 1.443168830000000034e-02
-9.743543930000000408e-01 1.085891540000000072e-06 1.096797820000000069e-02 2.498335310000000245e-04 1.443160689999999947e-02
-9.743550970000000033e-01 1.085892420000000088e-06 1.096799199999999933e-02 2.498334829999999994e-04 1.443159880000000062e-02
-9.743543949999999843e-01 1.085891530000000036e-06 1.096798539999999968e-02 2.498332210000000163e-04 1.443158259999999943e-02
-9.743545460000000435e-01 1.085891689999999981e-06 1.096798520000000048e-02 2.498333450000000087e-04 1.443159129999999936e-02
