--- tests/PseudoHomogeneous1DReactor/get_mass_fraction.asali
@@ -1,10 +1,0 @@
-1.000000000000000056e-01 4.000000000000000222e-01 0.000000000000000000e+00 5.000000000000000000e-01
-5.008586000000000288e-02 1.839680000000000005e-03 4.446109899999999837e-01 5.000000000000000000e-01
-5.008549999999999808e-02 1.837290000000000104e-03 4.446136199999999872e-01 5.000000000000000000e-01
-5.008514000000000022e-02 1.834899999999999986e-03 4.446162499999999906e-01 5.000000000000000000e-01
-5.008478000000000235e-02 1.832499999999999928e-03 4.446188099999999754e-01 5.000000000000000000e-01
-5.008441999999999755e-02 1.830100000000000086e-03 4.446207799999999932e-01 5.000000000000000000e-01
-5.008404999999999802e-02 1.827690000000000087e-03 4.446227500000000110e-01 5.000000000000000000e-01
-5.008369000000000015e-02 1.825280000000000088e-03 4.446247099999999786e-01 5.000000000000000000e-01
-5.008332000000000062e-02 1.822859999999999932e-03 4.446260699999999844e-01 5.000000000000000000e-01
-5.008296000000000275e-02 1.820439999999999992e-03 4.446271800000000107e-01 5.000000000000000000e-01
