--- tests/PseudoHomogeneous1DReactor/get_mole_fraction.asali
@@ -1,10 +1,0 @@
-6.647472599999999510e-01 1.675267799999999863e-01 0.000000000000000000e+00 1.677259700000000020e-01
-4.000842600000000249e-01 9.258600000000000293e-04 3.974411299999999758e-01 2.015487499999999987e-01
-4.000820800000000066e-01 9.246600000000000001e-04 3.974441599999999908e-01 2.015491000000000088e-01
-4.000798999999999883e-01 9.234599999999999710e-04 3.974472000000000005e-01 2.015494399999999964e-01
-4.000777400000000150e-01 9.222500000000000012e-04 3.974501999999999757e-01 2.015498000000000012e-01
-4.000757799999999920e-01 9.210499999999999720e-04 3.974528999999999979e-01 2.015502799999999983e-01
-4.000738200000000244e-01 9.198400000000000023e-04 3.974555900000000253e-01 2.015507500000000007e-01
-4.000718600000000014e-01 9.186300000000000325e-04 3.974582899999999919e-01 2.015512299999999979e-01
-4.000701099999999788e-01 9.174100000000000137e-04 3.974606600000000212e-01 2.015518199999999926e-01
-4.000684600000000146e-01 9.162000000000000439e-04 3.974629000000000079e-01 2.015524499999999941e-01
