--- tests/PseudoHomogeneous1DReactor/get_results.asali
@@ -1,10 +1,0 @@
-1.000000000000000056e-01 4.000000000000000222e-01 0.000000000000000000e+00 5.000000000000000000e-01 1.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 3.931499999999999773e+02
-5.008585730000000152e-02 1.839681440000000105e-03 4.446109859999999858e-01 5.000000000000000000e-01 9.743513350000000406e-01 1.086198959999999905e-06 1.097013569999999959e-02 2.497711849999999971e-04 1.442687460000000033e-02 3.041053960000000188e+03
-5.008549789999999780e-02 1.837289059999999957e-03 4.446136159999999893e-01 5.000000000000000000e-01 9.743518979999999940e-01 1.086344930000000015e-06 1.097118499999999962e-02 2.497417680000000141e-04 1.442463130000000079e-02 3.040862119999999777e+03
-5.008513860000000234e-02 1.834896670000000065e-03 4.446162470000000200e-01 5.000000000000000000e-01 9.743524609999999475e-01 1.086490899999999913e-06 1.097223420000000005e-02 2.497123499999999809e-04 1.442238789999999993e-02 3.040670290000000023e+03
-5.008477880000000021e-02 1.832502879999999912e-03 4.446188119999999744e-01 5.000000000000000000e-01 9.743531739999999886e-01 1.086637060000000077e-06 1.097328629999999930e-02 2.496829199999999957e-04 1.442014299999999985e-02 3.040478320000000167e+03
-5.008441520000000285e-02 1.830096270000000028e-03 4.446207790000000215e-01 5.000000000000000000e-01 9.743552450000000364e-01 1.086784889999999945e-06 1.097436400000000034e-02 2.496533729999999901e-04 1.441788429999999940e-02 3.040285100000000057e+03
-5.008405159999999856e-02 1.827689649999999968e-03 4.446227449999999859e-01 5.000000000000000000e-01 9.743573150000000016e-01 1.086932720000000026e-06 1.097544169999999965e-02 2.496238259999999844e-04 1.441562570000000029e-02 3.040091879999999946e+03
-5.008368800000000121e-02 1.825283040000000084e-03 4.446247119999999775e-01 5.000000000000000000e-01 9.743593860000000495e-01 1.087080549999999895e-06 1.097651950000000029e-02 2.495942789999999788e-04 1.441336699999999985e-02 3.039898659999999836e+03
-5.008332360000000011e-02 1.822863650000000092e-03 4.446260710000000116e-01 5.000000000000000000e-01 9.743626610000000188e-01 1.087229589999999918e-06 1.097761950000000070e-02 2.495646110000000230e-04 1.441109469999999997e-02 3.039704490000000078e+03
-5.008295890000000194e-02 1.820438929999999931e-03 4.446271760000000128e-01 5.000000000000000000e-01 9.743664389999999731e-01 1.087379140000000096e-06 1.097872880000000037e-02 2.495348919999999999e-04 1.440881679999999991e-02 3.039509919999999966e+03
