--- tests/PseudoHomogeneous1DReactor/get_temperature.asali
@@ -1,10 +1,0 @@
-3.931499999999999773e+02
-3.041053956700000072e+03
-3.040862122619999809e+03
-3.040670288530000107e+03
-3.040478317080000124e+03
-3.040285097689999930e+03
-3.040091878300000189e+03
-3.039898658919999889e+03
-3.039704487519999930e+03
-3.039509918570000082e+03
