+++ tests/SteadyStatePseudoHomogeneous1DReactor/example.asali
@@ -1,0 +1,10 @@
+1.000000000000000056e-01 4.000000000000000222e-01 0.000000000000000000e+00 5.000000000000000000e-01 1.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 5.131499999999999773e+02
+9.988359659999999918e-02 3.990762210000000088e-01 1.040182229999999904e-03 5.000000000000000000e-01 5.443995360000000033e-02 1.079237730000000045e-05 9.095796769999999754e-01 1.978531280000000081e-05 3.604647619999999664e-02 5.198489849999999706e+02
+9.974298439999999821e-02 3.979603210000000058e-01 2.296694499999999903e-03 5.000000000000000000e-01 6.058667580000000330e-02 1.538682330000000088e-05 9.050444279999999564e-01 2.106945289999999859e-05 3.428339710000000151e-02 5.279405169999999998e+02
+9.956705679999999781e-02 3.965641579999999999e-01 3.868785190000000127e-03 5.000000000000000000e-01 6.889717880000000017e-02 2.049195260000000133e-05 8.987813089999999727e-01 2.271401489999999912e-05 3.224034499999999664e-02 5.380632640000000038e+02
+9.933574900000000096e-02 3.947284959999999843e-01 5.935755180000000172e-03 5.000000000000000000e-01 8.085981059999999832e-02 2.624449789999999878e-05 8.892431450000000126e-01 2.492165559999999879e-05 2.978816100000000050e-02 5.513708880000000363e+02
+9.900724889999999723e-02 3.921215149999999761e-01 8.871235800000000157e-03 5.000000000000000000e-01 9.985078320000000118e-02 3.282377720000000272e-05 8.731641740000000151e-01 2.811159570000000126e-05 2.669475400000000109e-02 5.702670000000000528e+02
+9.847385549999999876e-02 3.878884989999999977e-01 1.363764540000000046e-02 5.000000000000000000e-01 1.355764409999999920e-01 4.036253049999999863e-05 8.418435680000000421e-01 3.333250490000000187e-05 2.248700530000000142e-02 6.009413540000000467e+02
+9.731821599999999906e-02 3.787173300000000187e-01 2.396445369999999908e-02 5.000000000000000000e-01 2.297079029999999911e-01 4.741591409999999768e-05 7.543844290000000230e-01 4.383790550000000091e-05 1.575921870000000086e-02 6.673678149999999505e+02
+9.205857039999999769e-02 3.369767209999999791e-01 7.096470900000000082e-02 5.000000000000000000e-01 6.509650630000000104e-01 2.894824620000000081e-05 3.437204979999999854e-01 5.963386550000000070e-05 3.952773009999999664e-03 9.685415719999999737e+02
+7.826868300000000556e-02 2.275400099999999870e-01 1.941913070000000074e-01 5.000000000000000000e-01 9.251537309999999792e-01 7.745410279999999590e-06 7.354771729999999563e-02 7.240431330000000442e-05 1.527528090000000076e-03 1.726351229999999987e+03
