+++ tests/SteadyStatePseudoHomogeneous1DReactor/get_coverage.asali
@@ -1,0 +1,10 @@
+1.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+5.009151209999999910e-02 7.002121240000000232e-06 9.124368830000000319e-01 1.883444219999999854e-05 3.744493589999999678e-02
+5.053058140000000337e-02 7.408818949999999611e-06 9.121508400000000183e-01 1.893206510000000042e-05 3.729656940000000004e-02
+5.097902790000000128e-02 7.818790669999999826e-06 9.118461920000000553e-01 1.903128940000000158e-05 3.714611629999999692e-02
+5.143856900000000343e-02 8.232653940000000175e-06 9.115382300000000049e-01 1.913260739999999931e-05 3.699403109999999800e-02
+5.190924329999999781e-02 8.650383630000000343e-06 9.112210120000000524e-01 1.923595509999999863e-05 3.684005679999999977e-02
+5.239072759999999912e-02 9.071404519999999542e-06 9.108954729999999556e-01 1.934124550000000157e-05 3.668442219999999698e-02
+5.288400099999999993e-02 9.496116420000000289e-06 9.105628980000000094e-01 1.944867859999999842e-05 3.652704040000000207e-02
+5.338861119999999821e-02 9.924553050000000390e-06 9.102054099999999925e-01 1.955797969999999855e-05 3.636711520000000114e-02
+5.390576150000000316e-02 1.035662539999999989e-05 9.098438469999999834e-01 1.966956650000000056e-05 3.620557989999999976e-02
