+++ tests/SteadyStatePseudoHomogeneous1DReactor/get_mass_fraction.asali
@@ -1,0 +1,10 @@
+1.000000000000000056e-01 4.000000000000000222e-01 0.000000000000000000e+00 5.000000000000000000e-01
+9.998922999999999839e-02 3.999145290000000186e-01 9.624064599999999383e-05 5.000000000000000000e-01
+9.997828709999999341e-02 3.998276869999999872e-01 1.940262869999999937e-04 5.000000000000000000e-01
+9.996716140000000417e-02 3.997393929999999984e-01 2.934459550000000127e-04 5.000000000000000000e-01
+9.995583569999999529e-02 3.996495119999999845e-01 3.946526100000000039e-04 5.000000000000000000e-01
+9.994430510000000545e-02 3.995580040000000221e-01 4.976904529999999927e-04 5.000000000000000000e-01
+9.993258180000000324e-02 3.994649690000000031e-01 6.024493260000000168e-04 5.000000000000000000e-01
+9.992064970000000268e-02 3.993702749999999968e-01 7.090747609999999623e-04 5.000000000000000000e-01
+9.990849870000000077e-02 3.992738450000000161e-01 8.176562129999999547e-04 5.000000000000000000e-01
+9.989613290000000367e-02 3.991757100000000169e-01 9.281574220000000323e-04 5.000000000000000000e-01
