+++ tests/SteadyStatePseudoHomogeneous1DReactor/get_mole_fraction.asali
@@ -1,0 +1,10 @@
+6.647472549999999814e-01 1.675267789999999868e-01 0.000000000000000000e+00 1.677259660000000041e-01
+6.646994560000000218e-01 1.674969779999999908e-01 7.159569179999999782e-05 1.677319700000000080e-01
+6.646508859999999963e-01 1.674666969999999977e-01 1.443459920000000082e-04 1.677380709999999886e-01
+6.646015010000000389e-01 1.674359069999999949e-01 2.183173979999999991e-04 1.677442749999999982e-01
+6.645512239999999959e-01 1.674045620000000067e-01 2.936239979999999872e-04 1.677505900000000050e-01
+6.645000340000000172e-01 1.673726470000000133e-01 3.702989889999999760e-04 1.677570200000000067e-01
+6.644479850000000187e-01 1.673401959999999966e-01 4.482606999999999889e-04 1.677635580000000071e-01
+6.643950039999999557e-01 1.673071650000000077e-01 5.276177490000000161e-04 1.677702130000000014e-01
+6.643410469999999624e-01 1.672735250000000062e-01 6.084370549999999723e-04 1.677769909999999864e-01
+6.642861310000000019e-01 1.672392880000000137e-01 6.906919779999999998e-04 1.677838889999999916e-01
