+++ tests/SteadyStatePseudoHomogeneous1DReactor/get_temperature.asali
@@ -1,0 +1,10 @@
+5.131499999999999773e+02
+5.137698302500000409e+02
+5.143996066500000097e+02
+5.150399022899999864e+02
+5.156917017700000088e+02
+5.163552895300000500e+02
+5.170299563600000283e+02
+5.177166389899999785e+02
+5.184159131899999693e+02
+5.191275454300000547e+02
