+++ tests/SteadyStatePseudoHomogeneous1DReactor/solve.asali
@@ -1,0 +1,10 @@
+1.000000000000000056e-01 4.000000000000000222e-01 0.000000000000000000e+00 5.000000000000000000e-01 1.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00 5.131499999999999773e+02
+9.998922999999999839e-02 3.999145290000000186e-01 9.624064599999999383e-05 5.000000000000000000e-01 5.009151209999999910e-02 7.002121240000000232e-06 9.124368830000000319e-01 1.883444219999999854e-05 3.744493589999999678e-02 5.137698299999999563e+02
+9.997828709999999341e-02 3.998276869999999872e-01 1.940262869999999937e-04 5.000000000000000000e-01 5.053058140000000337e-02 7.408818949999999611e-06 9.121508400000000183e-01 1.893206510000000042e-05 3.729656940000000004e-02 5.143996069999999463e+02
+9.996716140000000417e-02 3.997393929999999984e-01 2.934459550000000127e-04 5.000000000000000000e-01 5.097902790000000128e-02 7.818790669999999826e-06 9.118461920000000553e-01 1.903128940000000158e-05 3.714611629999999692e-02 5.150399019999999837e+02
+9.995583569999999529e-02 3.996495119999999845e-01 3.946526100000000039e-04 5.000000000000000000e-01 5.143856900000000343e-02 8.232653940000000175e-06 9.115382300000000049e-01 1.913260739999999931e-05 3.699403109999999800e-02 5.156917019999999638e+02
+9.994430510000000545e-02 3.995580040000000221e-01 4.976904529999999927e-04 5.000000000000000000e-01 5.190924329999999781e-02 8.650383630000000343e-06 9.112210120000000524e-01 1.923595509999999863e-05 3.684005679999999977e-02 5.163552899999999681e+02
+9.993258180000000324e-02 3.994649690000000031e-01 6.024493260000000168e-04 5.000000000000000000e-01 5.239072759999999912e-02 9.071404519999999542e-06 9.108954729999999556e-01 1.934124550000000157e-05 3.668442219999999698e-02 5.170299559999999701e+02
+9.992064970000000268e-02 3.993702749999999968e-01 7.090747609999999623e-04 5.000000000000000000e-01 5.288400099999999993e-02 9.496116420000000289e-06 9.105628980000000094e-01 1.944867859999999842e-05 3.652704040000000207e-02 5.177166389999999865e+02
+9.990849870000000077e-02 3.992738450000000161e-01 8.176562129999999547e-04 5.000000000000000000e-01 5.338861119999999821e-02 9.924553050000000390e-06 9.102054099999999925e-01 1.955797969999999855e-05 3.636711520000000114e-02 5.184159130000000459e+02
+9.989613290000000367e-02 3.991757100000000169e-01 9.281574220000000323e-04 5.000000000000000000e-01 5.390576150000000316e-02 1.035662539999999989e-05 9.098438469999999834e-01 1.966956650000000056e-05 3.620557989999999976e-02 5.191275450000000546e+02
